# add web exception handler

+++ global/sunny-cloud-framework/sunny-cloud-framework-core/src/main/java/com/sunny/cloud/framework/core/util/PageUtil.java
@@ -3,12 +3,10 @@ package com.sunny.cloud.framework.core.util;
 import com.github.pagehelper.PageInfo;
 import com.sunny.cloud.framework.core.model.PagingResult;
 
+import java.util.List;
+import java.util.function.Function;
+
 public class PageUtil {
-    /**
-     * 转换为pageResult
-     *
-     * @param page 分页结果
-     */
     public static <T> PagingResult<T> convert(PageInfo<T> page) {
         if (page == null) {
             return null;
@@ -21,21 +19,12 @@ public class PageUtil {
         return pagingResult;
     }
 
-    /**
-     * 转换分页结果;PO转DTO
-     *
-     * @param page
-     * @param convertListFunc
-     * @param <T>
-     * @param <S>
-     * @return
-     */
     public static <S, T> PagingResult<T> convert(PageInfo<S> page, Function<List<S>, List<T>> convertListFunc) {
         if (page == null) {
             return null;
         }
         PagingResult<T> pagingResult = new PagingResult<>();
-        pagingResult.setTotal(page.getTotal());
+        pagingResult.setTotal((int)page.getTotal());
         pagingResult.setList(convertListFunc.apply(page.getList()));
         pagingResult.setPage(page.getPageNum());
         pagingResult.setPageSize(page.getPageSize());
